refactor(ampd): remove unused Uint256
