+++ ampd/src/tm_address.rs
@@ -3,31 +3,12 @@ use std::hash::{Hash as StdHash, Hasher};
 
 use cosmrs::crypto;
 use cosmrs::AccountId;
-use cosmwasm_std::Uint256;
 use ethers::types::{Address, H256};
 use serde::{Deserialize, Serialize};
 
 pub type EVMAddress = Address;
 pub type Hash = H256;
 pub type PublicKey = crypto::PublicKey;
-
-#[derive(Deserialize, Debug)]
-#[serde(from = "Uint256")]
-pub struct U256(ethers::types::U256);
-
-impl From<Uint256> for U256 {
-    fn from(value: Uint256) -> Self {
-        Self(ethers::types::U256::from_big_endian(
-            value.to_be_bytes().as_slice(),
-        ))
-    }
-}
-
-impl AsRef<ethers::types::U256> for U256 {
-    fn as_ref(&self) -> &ethers::types::U256 {
-        &self.0
-    }
-}
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct TMAddress(AccountId);
@@ -61,7 +42,7 @@ pub mod test_utils {
     use ecdsa::SigningKey;
     use rand::rngs::OsRng;
 
-    use crate::types::{PublicKey, TMAddress};
+    use crate::tm_address::{PublicKey, TMAddress};
 
     impl TMAddress {
         pub fn random(prefix: &str) -> Self {
